+++ sn_registers/src/reg_crdt.rs
@@ -112,6 +112,12 @@ impl RegisterCrdt {
             .map(|(hash, node)| (EntryHash(hash), node.value.clone()))
             .collect()
     }
+
+    /// Access the underlying MerkleReg (e.g. for access to history)
+    /// NOTE: This API is unstable and may be removed in the future
+    pub(crate) fn merkle_reg(&self) -> &MerkleReg<Entry> {
+        &self.data
+    }
 }
 
 #[cfg(test)]